+++ examples/read_chunk_single_cell.rs
@@ -1,12 +1,14 @@
 use anyhow;
-use libradicl::rad_types::{self, RecordContext};
+use libradicl::chunk::Chunk;
+use libradicl::header;
+use libradicl::record::{AlevinFryReadRecord, AlevinFryRecordContext, RecordContext};
 use std::io::BufReader;
 
 fn main() -> anyhow::Result<()> {
     let fname = std::env::args().nth(1).expect("input filename");
     let f = std::fs::File::open(&fname)?;
     let mut ifile = BufReader::new(f);
-    let p = rad_types::RadPrelude::from_bytes(&mut ifile)?;
+    let p = header::RadPrelude::from_bytes(&mut ifile)?;
     if let Ok(summary) = p.summary(None) {
         println!("{}", summary);
     }
@@ -15,13 +17,12 @@ fn main() -> anyhow::Result<()> {
     println!("tag map {:?}\n", tag_map);
     // Any extra context we may need to parse the records. In this case, it's the
     // size of the barcode and the umi.
-    let tag_context = rad_types::AlevinFryRecordContext::get_context_from_tag_section(
+    let tag_context = AlevinFryRecordContext::get_context_from_tag_section(
         &p.file_tags,
         &p.read_tags,
         &p.aln_tags,
     )?;
-    let first_chunk =
-        rad_types::Chunk::<rad_types::AlevinFryReadRecord>::from_bytes(&mut ifile, &tag_context);
+    let first_chunk = Chunk::<AlevinFryReadRecord>::from_bytes(&mut ifile, &tag_context);
     println!(
         "Chunk :: nbytes: {}, nrecs: {}",
         first_chunk.nbytes, first_chunk.nrec
